Include Paragraph about OPTIONS-Call on client
As on posted on `https://forums.meteor.com/t/solved-meteor-http-client-sided-post-produces-options-call/30221/6` I improve the docs with warning about the client sided option-call if you use the same code as on the server and have a really basic rest-api where you need to implement the options-call yourself.

e.g. simplehttpserver in python has a really strange way of telling you that there is a problem as it tells you that the OPTIONS-Method is not defined. Because POST, GET, etc is also called method this mislead me into thinking that meteor does strange things here.