Switch to node 8, npm@5, and the official jsdoc.
Previously, we were using a fork of jsdoc in order to enable some newer,
experimental features we needed enabled in espree since its used to
parse our ES(YYYY+1) ECMAScript, but the official jsdoc has switched to
using babylon for its parser.  With that, we gain much of that out of
the box!