Change methodology for rendering sidebar sub-links.
Previously, the API box template was wrapping the "instance separator"
(that is to say the hash symbol in something like `String#toString` with
the HTML '<em>' tag in order to provide an italicised representation of
that symbol.  That styling seems to have been long lost and it's not
clear how much value it added anyway (I tried, and didn't even notice).

The value of this '<em>' wrapped "title" was being stored in the "title"
attribute (commonly used for a mouseover tooltip) of the H2 tag for each
section of the docs, though it was being improperly escaped so the actual HTML
was showing in the tooltip.  It's important to note that this "title"
attribute is used to add "sub-links" to the sidebar at runtime.

For an unspecified reason,
https://github.com/meteor/hexo-theme-meteor/commit/78ef41594675e1fdc9566904ed039d4f838f7d5c
added the escaping of this attribute when being used in the runtime
generation of links.  From the Meteor perspective, the content which was
now being escaped did not need to be escaped (it already was), and
therefore resulted HTMLesque titles being rendered into the sidebar.

As it's not clear what the reason for this commit was (I'm assuming
escaping is necessary for _some_ data coming into the sidebar, else the
commit wouldn't have been necessary), and given that I have a fairly clear
understanding of how the docs are instantiating this value, I'm opting to
remove the escaping of the title attribute (which actually _didn't_ need to
be escaped at all), but keep the escaping in place for the sub-link generation
in the event that it's necessary elsewhere.

It's situations like this where a verbose commit message can come in
quite handy.